backend/requirements.txt
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.2
Django==2.1.5
django-cors-headers==2.4.0
django-tastypie==0.14.2
django-webpack-loader==0.6.0
djangorestframework==3.9.1
idna==2.8
pkg-resources==0.0.0
python-dateutil==2.8.0
python-mimeparse==1.6.0
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
